requirements.txt
# Core dependencies
monai>=1.3.0
cucim>=23.10.0
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=9.5.0

# Data processing
pandas>=2.0.0
scikit-image>=0.21.0
scipy>=1.11.0

# Geometric operations
shapely>=2.0.0
geopandas>=0.14.0

# Visualization and debugging
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0

# Optional: GPU acceleration
# torch>=2.0.0
# torchvision>=0.15.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
